Function integration of new schema with battle room
