+++ server/src/rooms/schema/MyRoomState.ts
@@ -12,12 +12,3 @@ import { BattleTeam } from "./Group";
 export class MyRoomState extends Schema {
   @type({ map: Player }) players = new MapSchema<Player>();
 }
-
-export class BattleRoomState extends MyRoomState {
-  @type([BattleTeam]) teams = new ArraySchema<BattleTeam>();
-  @type({ map: Monster }) monsters = new MapSchema<Monster>();
-  @type("number") totalRounds: number;
-  @type("number") currentRound: number;
-  @type("number") roundTimeLeft: number;
-  @type("string") currentGameState: string;
-}
